+++ src/Traits/HelperTransformTrait.php
@@ -21,10 +21,11 @@ trait HelperTransformTrait
      */
     public static function transformToArray(mixed $value): array
     {
-        return $value = match (true) {
+        return match (true) {
             is_null($value) => [],
             $value instanceof BackedEnum => static::enumToArray($value),
             is_object($value) && method_exists($value, 'toArray') => $value->toArray(),
+            is_object($value) && method_exists($value, 'all') => $value->all(),
             is_array($value) => $value,
             is_scalar($value) => [$value],
             is_object($value) && is_callable($value) => static::transformToArray($value()),